feat: Add barretenberg wasm into `blackbox_solver`
