+++ acvm_js/Cargo.toml
@@ -28,17 +28,6 @@ gloo-utils = { version = "0.1", features = ["serde"] }
 js-sys = "0.3.62"
 const-str = "0.5.5"
 
-# Barretenberg WASM dependencies
-thiserror = "1.0.21"
-getrandom = { version = "0.2", features = [ "js" ]}
-rust-embed = { version = "6.6.0", features = [
-    "debug-embed",
-    "interpolate-folder-path",
-    "include-exclude",
-] }
-
-wasmer = { version = "3.3", default-features = false, features = [ "js-default" ] }
-
 [build-dependencies]
 build-data = "0.1.3"
 pkg-config = "0.3"
