+++ acvm_js/src/execute.rs
@@ -1,67 +1,14 @@
 use acvm::{
-    acir::{circuit::Circuit, BlackBoxFunc},
+    acir::circuit::Circuit,
     pwg::{ACVMStatus, ACVM},
-    BlackBoxFunctionSolver, BlackBoxResolutionError, FieldElement,
 };
 
 use wasm_bindgen::prelude::wasm_bindgen;
 
 use crate::{
-    barretenberg::{pedersen::Pedersen, scalar_mul::ScalarMul, schnorr::SchnorrSig, Barretenberg},
     foreign_call::{resolve_brillig, ForeignCallHandler},
     JsWitnessMap,
 };
-
-struct SimulatedBackend {
-    blackbox_vendor: Barretenberg,
-}
-
-impl SimulatedBackend {
-    async fn initialize() -> SimulatedBackend {
-        let blackbox_vendor = Barretenberg::new().await;
-        SimulatedBackend { blackbox_vendor }
-    }
-}
-
-impl BlackBoxFunctionSolver for SimulatedBackend {
-    fn schnorr_verify(
-        &self,
-        public_key_x: &FieldElement,
-        public_key_y: &FieldElement,
-        signature: &[u8],
-        message: &[u8],
-    ) -> Result<bool, BlackBoxResolutionError> {
-        let pub_key_bytes: Vec<u8> =
-            public_key_x.to_be_bytes().iter().copied().chain(public_key_y.to_be_bytes()).collect();
-
-        let pub_key: [u8; 64] = pub_key_bytes.try_into().unwrap();
-        let sig_s: [u8; 32] = signature[0..32].try_into().unwrap();
-        let sig_e: [u8; 32] = signature[32..64].try_into().unwrap();
-
-        self.blackbox_vendor.verify_signature(pub_key, sig_s, sig_e, message).map_err(|err| {
-            BlackBoxResolutionError::Failed(BlackBoxFunc::SchnorrVerify, err.to_string())
-        })
-    }
-
-    fn pedersen(
-        &self,
-        inputs: &[FieldElement],
-        domain_separator: u32,
-    ) -> Result<(FieldElement, FieldElement), BlackBoxResolutionError> {
-        self.blackbox_vendor
-            .encrypt(inputs.to_vec(), domain_separator)
-            .map_err(|err| BlackBoxResolutionError::Failed(BlackBoxFunc::Pedersen, err.to_string()))
-    }
-
-    fn fixed_base_scalar_mul(
-        &self,
-        input: &FieldElement,
-    ) -> Result<(FieldElement, FieldElement), BlackBoxResolutionError> {
-        self.blackbox_vendor.fixed_base(input).map_err(|err| {
-            BlackBoxResolutionError::Failed(BlackBoxFunc::FixedBaseScalarMul, err.to_string())
-        })
-    }
-}
 
 /// Executes an ACIR circuit to generate the solved witness from the initial witness.
 ///
@@ -78,7 +25,8 @@ pub async fn execute_circuit(
     console_error_panic_hook::set_once();
     let circuit: Circuit = Circuit::read(&*circuit).expect("Failed to deserialize circuit");
 
-    let backend = SimulatedBackend::initialize().await;
+    #[allow(deprecated)]
+    let backend = acvm::blackbox_solver::BarretenbergSolver::initialize().await;
     let mut acvm = ACVM::new(backend, circuit.opcodes, initial_witness.into());
 
     loop {
