+++ acvm_js/src/lib.rs
@@ -7,7 +7,6 @@
 
 cfg_if::cfg_if! {
     if #[cfg(target_arch = "wasm32")] {
-        mod barretenberg;
         mod build_info;
         mod compression;
         mod execute;
