+++ blackbox_solver/src/barretenberg/wasm/barretenberg_structures.rs
@@ -1,4 +1,4 @@
-use acvm::FieldElement;
+use acir::FieldElement;
 
 #[derive(Debug, Default)]
 pub(crate) struct Assignments(Vec<FieldElement>);
