+++ blackbox_solver/src/barretenberg/wasm/scalar_mul.rs
@@ -1,4 +1,4 @@
-use acvm::FieldElement;
+use acir::FieldElement;
 
 use super::{Barretenberg, Error, FIELD_BYTES};
